any_file_changed now uses globs instead of directory names

+++ spec/do_if_spec.rb
@@ -1,19 +1,19 @@
 require File.join(File.dirname(__FILE__), '../lib/do_if')
 
-def it_should_call_the_block(fixture_dir)
+def it_should_call_the_block(fixture_glob)
   it 'should call the block' do
     was_called = false
-    DoIf.any_file_changed fixture(fixture_dir) do
+    DoIf.any_file_changed fixture(fixture_glob) do
       was_called = true
     end
     was_called.should == true
   end
 end
 
-def it_should_not_call_the_block(fixture_dir)
+def it_should_not_call_the_block(fixture_glob)
   it 'should call the block' do
     was_called = false
-    DoIf.any_file_changed fixture(fixture_dir) do
+    DoIf.any_file_changed fixture(fixture_glob) do
       was_called = true
     end
     was_called.should == false
@@ -26,21 +26,21 @@ end
 
 describe DoIf do
   before do
-    FileUtils.rm(DoIf::YAML_FILE)
+    FileUtils.rm(DoIf::YAML_FILE) if File.exists?(DoIf::YAML_FILE)
   end
   
   describe '.any_file_changed' do
     describe 'when the specified directory hasnt been run before' do
-      it_should_call_the_block 'one_file'
+      it_should_call_the_block 'one_file/**/*'
     end
     
     describe 'when called multiple times on a single directory' do
       before do
-        DoIf.any_file_changed fixture('one_file') do end
+        DoIf.any_file_changed fixture('one_file/**/*') do end
       end
       
       describe 'when no files have changed since the last run' do
-        it_should_not_call_the_block 'one_file'
+        it_should_not_call_the_block 'one_file/**/*'
       end
       
       describe 'when a file has changed since the last run' do
@@ -48,7 +48,7 @@ describe DoIf do
           FileUtils.touch(fixture('one_file/1'))
         end
         
-        it_should_call_the_block 'one_file'
+        it_should_call_the_block 'one_file/**/*'
       end
 
       describe 'when a file has been added since the last run' do
@@ -60,17 +60,17 @@ describe DoIf do
           FileUtils.rm(fixture('many_files/42'))
         end
         
-        it_should_call_the_block 'many_files'
+        it_should_call_the_block 'many_files/**/*'
       end
       
       describe 'when a file has been deleted since the last run' do
         before do
           FileUtils.touch(fixture('many_files/42'))
-          DoIf.any_file_changed fixture('many_files') do end
+          DoIf.any_file_changed fixture('many_files/**/*') do end
           FileUtils.rm(fixture('many_files/42'))
         end
         
-        it_should_call_the_block 'many_files'
+        it_should_call_the_block 'many_files/**/*'
       end
     end
   end
